+++ src/components/account/ForgotForm.js
@@ -5,7 +5,7 @@ import TextField from 'material-ui/TextField';
 import Typography from 'material-ui/Typography';
 import {withStyles} from 'material-ui/styles';
 
-import config from '../../config/config.js';
+import config from '../../config.js';
 import AsylumConnectButton from '../AsylumConnectButton';
 
 const styles = theme => ({
