+++ src/components/account/GeneralSettings.js
@@ -13,7 +13,7 @@ import breakpoints from '../../theme/breakpoints';
 import Typography from 'material-ui/Typography';
 
 import 'whatwg-fetch';
-import config from '../../config/config.js';
+import config from '../../config.js';
 import {updateUserEmail, updateUserPassword} from '../../helpers/odasRequests';
 
 function TextMaskCustom(props) {
