+++ src/components/share/ShareFormContainer.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import 'whatwg-fetch';
 
-import config from '../../config/config.js';
+import config from '../../config.js';
 
 import {updateListPermissions} from '../../helpers/odasRequests';
 import ShareForm from './ShareForm';
