+++ src/helpers/odasRequests.js
@@ -1,4 +1,4 @@
-import config from '../config/config.js';
+import config from '../config.js';
 import 'whatwg-fetch';
 
 // One Degree Application Server
