+++ src/helpers/oneDegreePropertyMap.js
