+++ src/helpers/opportunities.js
